fix: monitor hid devices instead of input devices
And traverse the descendants tree when receiving a bound device.
This allows us to always find the same common ancestor to generate a consistent id for the devices
which fixes problems where a disconnected device has a different id than
the one we have.

Also remove a few unnecessary string clones and unnecessary transmutes