linux(drag n drop): Vastly improve drag and drop support
Previously we were only checking for XdndEnter, XdndDrop, XdndLeave Atoms.
We now reply to XdndPosition events with XdndStatus to let the source program know that we accepted or rejected the drop request. We also reply with XdndFinished to let the source program know we finished and clean up.
Also check for the XDND protocol version and reject enter requests for unknown versions.